% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamon.R
\name{hamon}
\alias{hamon}
\title{Hamon Potential Evapotranspiration Equation}
\usage{
hamon(par, tavg, lat, jday)
}
\arguments{
\item{par}{proportionality coefficient (unitless)}

\item{tavg}{vector of mean daily temperature (deg C)}

\item{lat}{latitude ()}

\item{jday}{a day number of the year (julian day of the year)}
}
\value{
outputs potential evapotranspiration (mm day-1)
}
\description{
The Hamon method is also considered as one of the simplest estimates 
    of potential Evapotranspiration.
}
\details{
For details see Haith and Shoemaker (1987)
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1deg
 }
}
}
