% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydroSim.R
\name{hydroSim}
\alias{hydroSim}
\title{Distributed Hydrology-Routing Model}
\usage{
hydroSim(par.hamon, par.snow17, par.sacsma, par.lohmann, tavg.grid, prcp.grid,
  lat.grid, elev.grid, area.grid, flength.grid, subcat.grid = NULL, jday)
}
\arguments{
\item{par.hamon}{parameters passed to hamon module}

\item{par.snow17}{parameters passed to snow17 module}

\item{par.sacsma}{parameters passed to sacsma module}

\item{par.lohmann}{parameters to be passed to lohmann module}

\item{tavg.grid}{a list of temperature series for each grid point}

\item{prcp.grid}{a list of precipitation series for each grid point}

\item{lat.grid}{a list of latitudes for each grid point}

\item{elev.grid}{a list of elevation information for each grid point}

\item{area.grid}{a list of area information for each grid point}

\item{flength.grid}{a list of flowlength information for each grid point}

\item{subcat.grid}{a list of subcatchment outlet points}

\item{jday}{a time-series of julian day of the year (1-3655)}
}
\value{
flow series at the specified outlet point(s)
}
\description{
To be completed...
}
\details{
to be completed...
}
