% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lohmann.R
\name{lohmann}
\alias{lohmann}
\title{River channel Routing model}
\usage{
lohmann(par, flength, uh.day = 96, ke = 12)
}
\arguments{
\item{par}{Vector of length four, describing the following model parameters (see details)}

\item{flength}{Flow length to the basin outlet (in meters)}

\item{uh.day}{Base time for river routing Unit Hydrograph (day) (Default: 96)}

\item{ke}{Base time for HRU unit hydrograph (day) (Default: 12)}
}
\value{
Placeholder
}
\description{
Mimics the routing model first described by Lohmann et al. (1998), which has been used within the 
  Variable Infiltration Capacity model (VIC)
}
\details{
Each grid cell is represented by a node in the channel network.
The total runoff and baseflow from each grid cell is first convolved with a unit hydrograph 
representing the distribution of travel times of water from its points of origin to the channel network
Then, each grid cell's input into the channel network is routed through the channel 
using linearized St. Venant's equations

The four parameters of the model are as follows:
NumRes: Grid Unit Hydrograph parameter defining the number of linear reservoirs
K: Grid Unit Hydrograph parameter defining the reservoir storage constant
VELO: wave velocity in the linearized Saint-Venant equation (cubic meters per sec)
DIFF: diffusivity in the linearized Saint-Venant equation (sq.meters per sec)

The detailed description of the routing can be found at:
Lohmann, D. E. Raschke, B. Nijssen and D.P. Lettenmaier, 1998: Regional scale hydrology: II. Application of the VIC-2L model to the Weser river, Germany, Hydrol. Sci. J., 43(1), 143-158.
}
