% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petHamon.R
\name{petHamon}
\alias{petHamon}
\title{Hamon Potential Evapotranspiration Equation}
\usage{
petHamon(coeff, tavg, lat, jday)
}
\arguments{
\item{coeff}{proportionality coefficient (unitless)}

\item{tavg}{vector of mean daily temperature (deg C)}

\item{lat}{latitude (deg)}

\item{jday}{a day number of the year (julian day of the year)}
}
\value{
outputs potential evapotranspiration (mm day-1)
}
\description{
The Hamon method is also considered as one of the simplest estimates 
 of Potential Evapotranspiration.
}
\details{
For details see Haith and Shoemaker (1987)
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
