% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lohamann.R
\name{lohamann}
\alias{lohamann}
\title{FUNCTION_TITLE}
\usage{
lohamann(par, inflow.direct, inflow.base, flowlen)
}
\arguments{
\item{flowlen}{PARAM_DESCRIPTION}

\item{UH_DAY}{PARAM_DESCRIPTION, Default: 96}

\item{KE}{PARAM_DESCRIPTION, Default: 12}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
