% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sacHydroSim.R
\name{sacHydroSim}
\alias{sacHydroSim}
\title{Distributed Hydrology-Routing Model}
\usage{
sacHydroSim(par_petHamon, par_snow17, par_sacsma, par_routLah, hru_tavg,
  hru_prcp, hru_lat, hru_elev, hru_area, hru_flowLength, hru_subcat = NULL,
  jday)
}
\arguments{
\item{par_petHamon}{PARAM_DESCRIPTION}

\item{par_snow17}{PARAM_DESCRIPTION}

\item{par_sacsma}{PARAM_DESCRIPTION}

\item{par_routLah}{PARAM_DESCRIPTION}

\item{hru_tavg}{PARAM_DESCRIPTION}

\item{hru_prcp}{PARAM_DESCRIPTION}

\item{hru_lat}{PARAM_DESCRIPTION}

\item{hru_elev}{PARAM_DESCRIPTION}

\item{hru_area}{PARAM_DESCRIPTION}

\item{hru_flowLength}{PARAM_DESCRIPTION}

\item{hru_subcat}{PARAM_DESCRIPTION}

\item{jday}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
See details
}
\details{
DETAILS
}
