% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sacsma.R
\name{sacSma}
\alias{sacSma}
\title{Sacremento Soil Moisture Accounting Model SAC-SMA}
\usage{
sacSma(par, ini.states = c(0, 0, 500, 500, 500, 0), prcp, pet)
}
\arguments{
\item{par}{model parameters}

\item{ini.states}{initial parameters}

\item{prcp}{daily precipitation data}

\item{pet}{potential evapotranspiration, in mm}

\item{verbose}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
