% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sacSma.R
\name{sacSma}
\alias{sacSma}
\title{Sacremento Soil Moisture Accounting Model SAC-SMA}
\usage{
sacSma(par, statesIni = c(0, 0, 5, 5, 5, 0), prcp, pet, lat, elev,
  verbose = FALSE)
}
\arguments{
\item{par}{a numeric vector of model parameters (see details)}

\item{statesIni}{initial parameters}

\item{prcp}{daily precipitation data}

\item{pet}{potential evapotranspiration, in mm}

\item{lat}{latitude information, in Deg}

\item{elev}{elevation, in meters}

\item{verbose}{additional model outputs (Default: FALSE)}
}
\value{
a numeric vector, defining surface flow (in mm)
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
