% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snow17.R
\name{snow17}
\alias{snow17}
\title{SNOW17 Model}
\usage{
snow17(par, prcp, tavg, elev, jdate, ini.states = c(0, 0, 0, 0))
}
\arguments{
\item{par}{snow module parameters}

\item{prcp}{numeric vector of precipitation time-series (mm)}

\item{tavg}{numeric vector of average temperatuer time-series (Deg C)}

\item{elev}{elevation (meters)}

\item{jdate}{julian day of the year}

\item{ini.states}{initial state parameters}

\item{verbose}{additional model outputs (default: FALSE)}
}
\value{
a numeric vector of precipitation series
}
\description{
Snow model
}
\details{
DETAILS
}
